Refactor SomlengRegion
