+++ app/types/region_type.rb
@@ -1,9 +1,9 @@
 class RegionType < ActiveRecord::Type::String
   def cast(value)
     return if value.blank?
-    return value if value.is_a?(SomlengRegions::Region)
+    return value if value.is_a?(SomlengRegion::Region)
 
-    SomlengRegions.regions.find_by(alias: value.to_s)
+    SomlengRegion::Region.find_by(alias: value.to_s)
   end
 
   def serialize(value)
