+++ config/application.rb
@@ -51,4 +51,4 @@ require "call_service"
 require "administrate_extensions"
 require "simple_form_components"
 require "twiml_parser"
-require "somleng_regions"
+require "somleng_region"
