+++ config/initializers/somleng_region.rb
@@ -1,4 +1,4 @@
-SomlengRegions.configure do |config|
+SomlengRegion.configure do |config|
   config.region_data = Rails.configuration.app_settings.fetch(:region_data)
   config.stub_regions = Rails.configuration.app_settings.fetch(:stub_regions)
 end
