+++ lib/somleng_region/configuration.rb
@@ -1,4 +1,4 @@
-module SomlengRegions
+module SomlengRegion
   class Configuration
     attr_accessor :region_data, :stub_regions
   end
