--- lib/somleng_regions/region.rb
@@ -1,5 +1,0 @@
-require "ostruct"
-
-module SomlengRegions
-  class Region < OpenStruct; end
-end
